+++ packages/editor/src/bindings/index.js
@@ -1,7 +1,7 @@
 /**
  * WordPress dependencies
  */
-import { store as blockEditorStore } from '@wordpress/block-editor';
+import { store as blocksStore } from '@wordpress/blocks';
 import { dispatch } from '@wordpress/data';
 /**
  * Internal dependencies
@@ -10,6 +10,6 @@ import { unlock } from '../lock-unlock';
 import patternOverrides from './pattern-overrides';
 import postMeta from './post-meta';
 
-const { registerBlockBindingsSource } = unlock( dispatch( blockEditorStore ) );
+const { registerBlockBindingsSource } = unlock( dispatch( blocksStore ) );
 registerBlockBindingsSource( patternOverrides );
 registerBlockBindingsSource( postMeta );
